+++ src/components/FileInput/FileInput.spec.tsx
@@ -1,0 +1,30 @@
+import { fireEvent, render, waitFor } from '@testing-library/react';
+
+import FileInput from './FileInput';
+
+describe('FileInput component', () => {
+  const onClickMock = jest.fn();
+  const mockProps = {
+    id: 'test-id',
+    testId: 'test-file-input',
+    onClick: onClickMock,
+    multiple: true,
+  };
+  const { getByTestId } = render(<FileInput {...mockProps} />);
+  it('should render FileInput component with required props', () => {
+    waitFor(() => {
+      expect(getByTestId('test-file-input')).toBeInTheDocument();
+    });
+  });
+  it('should handle onClick event', () => {
+    waitFor(() => {
+      fireEvent.click(getByTestId('test-file-input'));
+      expect(onClickMock).toHaveBeenCalled();
+    });
+  });
+  it('should render with multiple', () => {
+    waitFor(() => {
+      expect(getByTestId('test-file-input')).toHaveAttribute('multiple');
+    });
+  });
+});
